notion-sync/requirements.txt
notion-client==2.2.1
python-dotenv==1.0.0
watchdog==3.0.0
click==8.1.7
rich==13.7.0 